feat(summary): AZ function Task owner report sync (#713)
- [x] New feature
- [ ] Bug fix
- [ ] High impact

**Description of work:**
<!--- Please give a description of the work --->


[AB#57449](https://statoil-proview.visualstudio.com/787035c2-8cf2-4d73-a83e-bb0e6d937eec/_workitems/edit/57449)

Similar to the resource owners sync function, this retrieves recipients
(admins and project director) and project data from org api. Then stores
it in the Summary api and sends it to the queue to be processed by a
worker.

- Only projects that are active are synced
- Only projects that have an active project director or have any direct
admins will be synced.

**Testing:**
- [x] Can be tested
- [ ] Automatic tests created / updated
- [x] Local tests are passing

<!--- Please give a description of how this can be tested --->

I've ran this locally and in azure pr env against this other PR env:
#711

Will run infra pipeline to create pr/ci/fqa queues and az funcs when
this has been code reviewed. Done!

**Checklist:**
- [x] Considered automated tests
- [x] Considered updating specification / documentation
- [x] Considered work items 
- [x] Considered security
- [x] Performed developer testing
- [x] Checklist finalized / ready for review

<!--- Other comments --->