feat(summary): Add Task Owner Report endpoints (#711)
- [x] New feature
- [ ] Bug fix
- [ ] High impact

**Description of work:**
<!--- Please give a description of the work --->


[AB#57125](https://statoil-proview.visualstudio.com/787035c2-8cf2-4d73-a83e-bb0e6d937eec/_workitems/edit/57125)

- Adds necessary endpoints for task owner reports. This includes get and
put for projects and task owner reports
- Removed personIdentifiers validation as I believe it is not the
summary apis responsibility to verify this. In the case there is an
invalid person, then the whole put request errors out. This in turn may
lead to an entire project or department not reviving a report.

Code is quite simple, but there is a lot of boilerplate code that needs
to be written

**Testing:**
- [x] Can be tested
- [x] Automatic tests created / updated
- [x] Local tests are passing

<!--- Please give a description of how this can be tested --->

Written simple tests. Api does nothing by itself and a larger test will
be conducted once azure functions are up and running

**Checklist:**
- [x] Considered automated tests
- [x] Considered updating specification / documentation
- [x] Considered work items 
- [x] Considered security
- [x] Performed developer testing
- [x] Checklist finalized / ready for review

<!--- Other comments --->