chore(iac): Disable resources connection string for PR env instead of soft deleting it (#719)
- [ ] New feature
- [ ] Bug fix
- [ ] High impact

**Description of work:**
<!--- Please give a description of the work --->

When deploying infra IAC the PR env would fail because it would soft
delete the secret but not purge it. With this PR it will instead be
disabled.

Example of a pipeline run that fails

https://statoil-proview.visualstudio.com/Fusion%20Resource%20Allocation/_build/results?buildId=131253&view=logs&j=013c5867-0d4e-5d8f-4115-5efe4ac203c0&t=bc5fcbcb-7453-5c1d-5540-a7bbd223bfa7&s=2741fbd7-98d1-5541-d761-9bcc81bce3d2

I've circumvented this earlier by **personally** purging the secret
myself 😡

**Testing:**
- [ ] Can be tested
- [ ] Automatic tests created / updated
- [ ] Local tests are passing

<!--- Please give a description of how this can be tested --->

Ran this pipeline and created a PR that ran resource api. Then tested
that it could talk to the DB. TEST PR #720

**Checklist:**
- [ ] Considered automated tests
- [ ] Considered updating specification / documentation
- [ ] Considered work items 
- [ ] Considered security
- [x] Performed developer testing
- [x] Checklist finalized / ready for review

<!--- Other comments --->