feat: Support workday manager structure (#721)
- [x] New feature
- [ ] Bug fix
- [ ] High impact

**Description of work:**
Managers structure is changed when workday is rolled out. The manager will no longer be in the department he/she is maanger for. This means they are moved one level up.

Logics we have where `isResourceOwner` is combined with `fullDepartment` cannot be used as is.


**Testing:**
- [ ] Can be tested
- [ ] Automatic tests created / updated
- [ ] Local tests are passing

TBD, tests must be refactored as setup for manager structure will be
more complex


**Checklist:**
- [ ] Considered automated tests
- [ ] Considered updating specification / documentation
- [ ] Considered work items 
- [ ] Considered security
- [ ] Performed developer testing
- [ ] Checklist finalized / ready for review

---------

Co-authored-by: Jonathan Idland Olsnes <73334350+Jonathanio123@users.noreply.github.com>