fix: set isResourceOwner to true if user has any department responsib… (#737)
- [ ] New feature
- [x] Bug fix
- [ ] High impact

**Description of work:**
<!--- Please give a description of the work --->
Because of the new structure introduces with Workday, a user is not in
the same department as the department the user has responsibility of.
This causes the isResourceOwner property to not get set properly.

In addition to cover all cases when a user isResourceOwner, the property
is set to true when the user has any department with responsibility


[AB#59299](https://statoil-proview.visualstudio.com/787035c2-8cf2-4d73-a83e-bb0e6d937eec/_workitems/edit/59299)

**Testing:**
- [ ] Can be tested
- [ ] Automatic tests created / updated
- [ ] Local tests are passing

<!--- Please give a description of how this can be tested --->


**Checklist:**
- [ ] Considered automated tests
- [ ] Considered updating specification / documentation
- [ ] Considered work items 
- [ ] Considered security
- [ ] Performed developer testing
- [ ] Checklist finalized / ready for review

<!--- Other comments --->