Added new endpoints for Analytics (#745)
- [x] New feature
- [ ] Bug fix
- [ ] High impact

**Description of work:**

Thru the old endpoints PowerBI were able to pull alot of data and this
forced us to upgrade the technical requirements of the server. To
mitigate this problem new version of the same endpoints has been created
using IAsyncEnumerable to return the data.

Further improvements might be needed.


[AB#60374](https://statoil-proview.visualstudio.com/787035c2-8cf2-4d73-a83e-bb0e6d937eec/_workitems/edit/60374)

Some test results for /analytics/requests/internal v1 vs v2:

In CI, both return 2098 enteties and 3.1 MB of data. V2 is 100-200 ms
slower. Using Visual Studio diagnostics tools the object difference is
170 000 for v1 and 1686 for v2. The heap difference for v1 is 16 000 KB
and for v2 is -45 KB. This measurement is tkane with a snapshot before
and after execution of endpoint.

**Testing:**
- [x] Can be tested
- [ ] Automatic tests created / updated
- [ ] Local tests are passing

<!--- Please give a description of how this can be tested --->


**Checklist:**
- [x] Considered automated tests
- [ ] Considered updating specification / documentation
- [x] Considered work items 
- [x] Considered security
- [x] Performed developer testing
- [x] Checklist finalized / ready for review

<!--- Other comments --->