feat: Fetch org unit personnel based on sap id (#747)
- [x] New feature
- [x] Bug fix
- [ ] High impact

**Description of work:**
Currently we are fetching org unit personnel based on manager in the
search index. This works most of the time, however is not a correct
picture.
Core team has added sap id to the search index, so we can now query on
this attribute instead. Should give us a more correct answer, regardless
of who is manager/acting etc etc.

Risk: Could give a different picture, so where the "incorrect result"
has become a "feature", it could lead to error reports.


**Testing:**
- [x] Can be tested
- [ ] Automatic tests created / updated
- [ ] Local tests are passing

Verified by comparing selection of responses from ci and pr api.

**Checklist:**
- [ ] Considered automated tests
- [ ] Considered updating specification / documentation
- [ ] Considered work items 
- [ ] Considered security
- [ ] Performed developer testing
- [ ] Checklist finalized / ready for review