fix: only return existing departments when fetching responsibilities for profile.  (#746)
- [ ] New feature
- [x] Bug fix
- [ ] High impact

**Description of work:**
<!--- Please give a description of the work --->
Work Item:
[AB#60494](https://statoil-proview.visualstudio.com/787035c2-8cf2-4d73-a83e-bb0e6d937eec/_workitems/edit/60494)

Added a check to ensure departments exist before adding to list of
delegated responsibilities. This issue occurred because a department was
removed with active delegated responsibles without removing them as
well. This will ensure the api returns valid data, but the PIMS sync
should remove the delegated responsibles.

**Testing:**
- [ ] Can be tested
- [ ] Automatic tests created / updated
- [x] Local tests are passing

<!--- Please give a description of how this can be tested --->
Unable to test with the current mock of the LineOrgService. There is no
ability to remove departments. The existing tests do confirm that it is
working correctly under normal circumstances. It has been/can be tested
in swagger with the user that reported the bug.

**Checklist:**
- [x] Considered automated tests
- [x] Considered updating specification / documentation
- [x] Considered work items 
- [x] Considered security
- [x] Performed developer testing
- [x] Checklist finalized / ready for review

<!--- Other comments --->