feat: require location in validation for other tasks (#751)
- [ ] New feature
- [x] Bug fix
- [ ] High impact

**Description of work:**

Added location requirement to validation of CreatePersonAbsenceRequest
and UpdatePersonAbsenceRequest for absences of type `OtherTasks`.

Added the method `LocationWillBeNull` to the
`InternalRequestsController` class to check that a location will be set
for the following endpoints:
- [HttpPost("/departments/{departmentString}/resources/requests")]
- [HttpPatch("/resources/requests/internal/{requestId}")]
-
[HttpPatch("/departments/{departmentString}/resources/requests/{requestId}")]

The logic is a little complicated, it needs to ensure that after changes
will have been applied the request will have a valid location. It needs
to take into account whether the request already has a position, if a
location is being proposed, and that the proposed location isn't null,
i.e. removing an existing location.

Tests were updated to have a location by default to not cause errors in
other existing tests, and tests for the changes were added.

Work Item:
[AB#61121](https://statoil-proview.visualstudio.com/787035c2-8cf2-4d73-a83e-bb0e6d937eec/_workitems/edit/61121)

See frontend pr:
https://github.com/equinor/fusion-resource-allocation-apps/pull/897

**Testing:**
- [x] Can be tested
- [x] Automatic tests created / updated
- [x] Local tests are passing

<!--- Please give a description of how this can be tested --->


**Checklist:**
- [x] Considered automated tests
- [ ] Considered updating specification / documentation
- [x] Considered work items 
- [x] Considered security
- [x] Performed developer testing
- [x] Checklist finalized / ready for review

<!--- Other comments --->