fix: api should not return null for project state (#761)
- [ ] New feature
- [x] Bug fix
- [ ] High impact

**Description of work:**
<!--- Please give a description of the work --->

Link to work item:
[AB#60951](https://statoil-proview.visualstudio.com/787035c2-8cf2-4d73-a83e-bb0e6d937eec/_workitems/edit/60951)

Multi-pronged approach setting the project state to `"ACTIVE"` if it is
null. This is done in the `QueryProject` and `QueryProjectRef`
constructors when querying projects, when projects are added to the
database, and when projects are synced with org. Over time all null
values should be replaced in the database, and the check in the query
layer will act as a failsafe.

Created a string extension to set the state correctly, this will also
allow for adding checks in the future.

**Testing:**
- [x] Can be tested
- [x] Automatic tests created / updated
- [x] Local tests are passing

<!--- Please give a description of how this can be tested --->


**Checklist:**
- [x] Considered automated tests
- [x] Considered updating specification / documentation
- [x] Considered work items 
- [x] Considered security
- [x] Performed developer testing
- [x] Checklist finalized / ready for review

<!--- Other comments --->