fix(resources): Allow location to be null when patching a request (#772)
- [ ] New feature
- [x] Bug fix
- [ ] High impact

**Description of work:**
<!--- Please give a description of the work --->

[AB#61687](https://statoil-proview.visualstudio.com/787035c2-8cf2-4d73-a83e-bb0e6d937eec/_workitems/edit/61687)

Seems it was a bit too strict to check for location when patching a
request. Now instead of ensuring a location is set when patching a
request, it is instead ensured when the request is proposed from the
resource owner side.

As with many of these endpoints, there is some business logic placed in
the frontend and backend...

**Testing:**
- [x] Can be tested
- [x] Automatic tests created / updated
- [x] Local tests are passing

<!--- Please give a description of how this can be tested --->


**Checklist:**
- [x] Considered automated tests
- [x] Considered updating specification / documentation
- [x] Considered work items 
- [x] Considered security
- [x] Performed developer testing
- [x] Checklist finalized / ready for review

<!--- Other comments --->