fix: check all responsible departments for delegate access (#785)
- [ ] New feature
- [x] Bug fix
- [ ] High impact

**Description of work:**
<!--- Please give a description of the work --->

Work Item:
[AB#63280](https://statoil-proview.visualstudio.com/787035c2-8cf2-4d73-a83e-bb0e6d937eec/_workitems/edit/63280)

The authorization check for delegating access to a department has
previously only checked the first resource owner claim for the user.
Updated the check to compare against all resource owner claims.

**Testing:**
- [x] Can be tested
- [x] Automatic tests created / updated
- [x] Local tests are passing

<!--- Please give a description of how this can be tested --->
I was able to reproduce the bug by adding myself as a resource owner to
multiple departments via ci lineorg. It is important that the
departments have different paths (parents will have access to delegate
responsibility for children). Personnel-Allocation running in localhost
can point to the pr to verify it.

**Checklist:**
- [ ] Considered automated tests
- [ ] Considered updating specification / documentation
- [ ] Considered work items 
- [ ] Considered security
- [ ] Performed developer testing
- [ ] Checklist finalized / ready for review

<!--- Other comments --->