fix(resources): remove future allocation (#796)
- [ ] New feature
- [x] Bug fix
- [ ] High impact

**Description of work:**
<!--- Please give a description of the work --->

Note: 3 tests are currently failing because of the two spots where
publishing events are commented out. After PR testing is complete and
the comments have been removed, all tests should run successfully before
merging.

Work Item:
[AB#57209](https://statoil-proview.visualstudio.com/787035c2-8cf2-4d73-a83e-bb0e6d937eec/_workitems/edit/57209)

- Provisioning resource owner requests only checked dates, not time.
When removing a person from the instance it would create 2 splits, the
first up until 00:00 on the last day of the allocation, and another from
00:00-23:59:59 on the last day. Updated to include the time of the
allocation.

- If the request sub-type is "RemoveResource", explicitly set the
proposed person to null. It was previously being ignored when null,
which is why the person was never removed.

**Testing:**
- [x] Can be tested
- [x] Automatic tests created / updated
- [x] Local tests are passing

<!--- Please give a description of how this can be tested --->
org-admin pr link
https://1377-org-admin.pr.app.fusion-dev.net/apps/org-admin
personnel-allocation pr link
https://1377-personnel-allocation.pr.app.fusion-dev.net/apps/personnel-allocation

- Using the org-admin pr link create a future split, or find an existing
one to use for testing.
- Allocate someone to the split
- Using the personnel-allocation pr link find the person you allocated
to the split
- Find the future allocation, click Edit allocation, End person
allocation, Request change (no need to update any fields)
- Back in org-admin go to requests and find the Remove Resource request
- Accept the change

The request is now stuck in provisioning because the event publishing
here has been commented out. If this hadn't been commented out the
resource function would trigger provisioning on the ci resources-api,
not the one in this pr. So provisioning has to be executed manually. The
redirect for swagger is not working correctly, but a curl command can be
executed instead.

- Run the following curl command, replacing BEARER-TOKEN and REQUEST-ID
with the relevant values.
```bash
curl -I -X 'POST' \
  'https://fra-resources-796.pr.api.fusion-dev.net/resources/requests/internal/REQUEST-ID/provision?force=false&api-version=1.0' \
  -H 'accept: text/plain' \
  -H 'Authorization: Bearer BEARER-TOKEN'

```
- The split should now have no-one allocated.


**Checklist:**
- [x] Considered automated tests
- [x] Considered updating specification / documentation
- [x] Considered work items 
- [x] Considered security
- [x] Performed developer testing
- [x] Checklist finalized / ready for review

<!--- Other comments --->